sddm: bump revision
Install the latest revision from the 'develop' branch.

This fixes an issue I had with the 0.21.0 release.

More specifically, in the latest revision, the PAM files have been
removed, distros are now responsible for providing the necessary PAM
files in /etc/pam.d.

This fixes an issue I had wrt to the sddm-greeter file which defaulted
to the debian ones (probably a Yocto installation issue) and was
causing a black screen when logging out of the session.

Note: the sddm-greeter file comes from the Arch PAM modules that
previously shipped with sddm.

Note: I removed the 0001-Workaround-missing-sessions.patch patch since
I do not think we are affected by this issue: all wayland session
files ship with the full path and the TryExec variable is set
correctly. I also searched sddm's issues on GH and did not find
anything related to this issue so my guess is that we can safely drop
it.

Signed-off-by: Marc Ferland <marc.ferland@sonatest.com>