systemd: always enable xdg-autostart option with kde
Enabling the xdg-autostart option in systemd in recommended so that
plasma and all related services can be started (and stopped) with
systemd instead of relying on what is called the 'classic boot' model.

This fixes an important issue for me where the platofrm would freeze
for 90 seconds on restart or shutdown because of the kded6 daemon not
quitting.

See the following for an example:

https://bugs.kde.org/show_bug.cgi?id=478313

As an added bonus, the sddm.service file can now be left as-is since
it will get started automatically (it was already part of the graphics
target).

Signed-off-by: Marc Ferland <marc.ferland@sonatest.com>